Move inventory events, remove unnecessary imports
