+++ src/main/java/nl/mtvehicles/core/infrastructure/helpers/MenuUtils.java
@@ -3,8 +3,7 @@ package nl.mtvehicles.core.infrastructure.helpers;
 import nl.mtvehicles.core.infrastructure.models.ConfigUtils;
 import nl.mtvehicles.core.infrastructure.models.Vehicle;
 import nl.mtvehicles.core.infrastructure.modules.ConfigModule;
-import nl.mtvehicles.core.inventory.InventoryClickEvent;
-import nl.mtvehicles.core.Main;
+import nl.mtvehicles.core.events.inventory.InventoryClickEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
